+++ .gitignore
@@ -31,3 +31,5 @@ config.js
 /client/babelified/*
 /client/dist/bundle.js
 /client/dist/bundle.js.map
+
+.DS_Store
