+++ server/middlewares/drinks.js
@@ -5,10 +5,6 @@ var router = express.Router();
 var request = require('supertest');
 var app = express();
 
-// we want to respond to post request with a list of all rated drinks
-  // '/rateDrink'
-  // '/getRecommendations'
-
 /*
 { 
   user: token123124123123123122,
@@ -21,32 +17,20 @@ var app = express();
 /* GET home page. */
 app.get('/drinks', function(req, res, next) {
   //respond with data of all drinks in db
-   console.log(req.body);
-   drinks.getAllDrinks(function(results) {
-
+  // console.log(req.body);
+  drinks.getAllDrinks(function(results) {
     // console.log('results from router.get callback', results);
-    //console.log(results);
-     var testJSON = {'results':results};
-     // console.log('THIS IS THE JSON');
-     // console.log(testJSON);
-     res.json(testJSON);
-
-     //res.send({ message: 'hey' });
+    // console.log(results);
+    var testJSON = {'results':results};
+    res.json(testJSON);
   });
-  //console.log(res);
-  //console.log('adf');
 });
 
-//test
-
-
-app.get('/randomDrinks', function(req, res, next) {
+app.get('/drinks/randomDrinks', function(req, res, next) {
   //respond with data of all drinks in db
    drinks.getRandomDrinks(function(results) {
     // console.log('results length from router.get callback', results);
-    
     res.send(results);
-
   });
 });
 
@@ -57,7 +41,7 @@ request(app)
   .end(function(err, res){
     if (err) throw err;
 
-     console.log(res.body.results);
+     // console.log(res);
   });
 
 module.exports = app;
